fix: react suppressHydrationWarning (#59)
This script is being placed in the head of my application it adds the `data-pt-script` attribute to the script tag *BEFORE* React has a chance to hydrate. Because the attribute has been added before react has had a chance to hydrate, React reports a warning to the console `Warning: Extra attributes from the server: data-pt-script`. 

Adding `suppressHydrationWarning` to the script tag will cause React to ignore that warning.