We need to include an extra element for attribution config
This is needed because Astro is apparently "bundling" scripts/modules
itself, so the original `script` tag (and all of its attributes) gets
lost in the process. By adding an extra element (that is not a `script`)
with the correct `id` attribute we can use it to configure the
attribution behavior.