partytownVite does not copy resource in static dir
Hello.

I followed the above steps, but when I deploy the solution to production. I receive a 404 when the script attempts to retrieve `partytown-sw.js`. The culprit seems to be `partytownVite` because if I manually copy `partytown-sw.js` to `/static/~partytown/` the bug is resolved.  See https://github.com/cesarnml/partytown-bug for a minimal reproduction.